vision/vision_general/requirements.txt
opencv-python
pathlib
tqdm
numpy<2
face_recognition
ultralytics
moondream==0.0.5
pillow
# pybind11>=2.12